backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
web3==6.11.1
eth-account==0.8.0
tweepy==4.14.0
discord.py==2.3.2
python-multipart==0.0.6
pydantic==2.4.2
alembic==1.12.1 